requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-multipart
trafilatura
networkx
beautifulsoup4
requests
python-dotenv
wikipedia-api
googlesearch-python
